refactor: service 인터페이스 삭제
